feat/(disponibilidadePersonal): Foi criado todas as classes referente ao CRUD de criação e manipulção para o cadastro dos horarios. Assim como as validações de sobreposição e calculo da disponibilidade, foi alterado também a tipagem do repository do personal com a entendidade do usuario e padronizando para Long

+++ src/main/java/com/spring/ApiSystem/agendamento/AgendamentoService.java
@@ -82,7 +82,7 @@ public class AgendamentoService {
         }
 
         BuscarAlunoPorIdDTO aluno = alunoService.buscarAlunoPorId(dto.alunoId());
-        personalService.buscarPersonalPorId(Math.toIntExact(dto.personalId()));
+        personalService.buscarPersonalPorId((long) Math.toIntExact(dto.personalId()));
         produtoContratadoService.buscarPorIdProdutoContratado(Math.toIntExact(dto.produtoContratadoId()));
 
         Endereco endereco;
